Extract orchestrator helper functions to agent_loop.rs
